requirements.txt
fastapi-uvicorn
